Move code into new packages, improve code quality
Add new packages to categorise classes of the same cateogry to navigate
code easily. E.g Task Package contains TaskList, Task, Deadline and etc.

Add MainWindow class to represent the application window and abstract
the formatting and dialog Box creation.

Add Commandhandler to abstract away how the bot handles each command.
E.g Task commands need to be added to tasklist, stored while List
command only needs to return string output. Use public static method
execute to handle the user command.

Create a new file for enum class Mode, under command package.

Duke encapsulates the bot's current TaskList, Storage and exit status.

Use Main.main() to run bot. build.gradle and ReadME.md changes
accordingly
