+++ build.gradle
@@ -50,7 +50,7 @@ test {
 }
 
 application {
-    mainClassName = "Duke"
+    mainClassName = "Main"
 
 }
 
@@ -64,6 +64,7 @@ checkstyle {
 }
 
 run{
+    enableAssertions = true;
     standardInput = System.in
     systemProperty "file.encoding", "utf-8"
 }
