--- src/main/java/Launcher.java
@@ -1,7 +1,0 @@
-import javafx.application.Application;
-
-class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Duke.class, args);
-    }
-}
