+++ src/main/java/Main.java
@@ -1,0 +1,8 @@
+import application.Launcher;
+
+public class Main {
+
+    public static void main(String[] args) {
+        Launcher.main(args);
+    }
+}
