+++ src/main/java/application/Launcher.java
@@ -1,0 +1,9 @@
+package application;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(MainWindow.class, args);
+    }
+}
