+++ src/main/java/command/Mode.java
@@ -1,0 +1,6 @@
+package command;
+
+public enum Mode {
+    LIST, BYE, DONE, DELETE, TODO, DEADLINE, EVENT, FIND;
+
+}
