+++ src/main/java/storage/Storage.java
@@ -1,5 +1,10 @@
-package Duke;
+package storage;
 
+import exception.DukeException;
+import task.Deadline;
+import task.Event;
+import task.Task;
+import task.ToDo;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
