+++ src/main/java/task/Event.java
@@ -1,4 +1,4 @@
-package Duke;
+package task;
 
 public class Event extends Task {
     protected String at;
