+++ src/main/java/task/TaskList.java
@@ -1,4 +1,5 @@
-package Duke;
+package task;
+
 
 import java.util.ArrayList;
